pyproject.toml
[project]
name = "convertdocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "img2pdf>=0.6.1",
    "pymupdf>=1.26.3",
    "python-pptx>=0.6.23",
    "pillow>=10.5.0",
]
